% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{nhgCI_M}
\alias{nhgCI_M}
\title{Compute Confidence Intervals for M (Unknown Successes)}
\usage{
nhgCI_M(N, m, conf_level = 0.95, method = "Analog to Clopper-Pearson")
}
\arguments{
\item{N}{Total population size.}

\item{m}{Fixed number of successes to be observed.}

\item{conf_level}{Confidence level for the interval (default = 0.95).}

\item{method}{Method to compute the interval. Choices are "Analog to Clopper-Pearson", "CMC", "CG", or "Blaker".}
}
\value{
A data frame containing the confidence intervals for each observed number of failures observed before the mth success (x).
}
\description{
Wrapper function to compute confidence intervals for the number of successes (M) in the Negative Hypergeometric Distribution using different methods.
}
\note{
The function assumes \code{m \geq 1} and \code{m \leq N}.
Confidence intervals are computed for all appropriate values of the number of observed failures (\code{x}) based on the specified \code{N} and \code{m}.
The full set of valid \code{x} values is output without the need for additional parameter tuning.
}
\examples{
# Example 1: Using the Analog to Clopper-Pearson method
nhgCI_M(N = 30, m = 5, conf_level = 0.95, method = "Analog to Clopper-Pearson")

# Example 2: Using the CMC method
nhgCI_M(N = 20, m = 4, conf_level = 0.90, method = "CMC")

# Example 3: Using the CG method
nhgCI_M(N = 15, m = 3, conf_level = 0.99, method = "CG")

# Example 4: Using the Blaker method
nhgCI_M(N = 25, m = 7, conf_level = 0.95, method = "Blaker")

}
