% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{nhgCI_N}
\alias{nhgCI_N}
\title{Compute Confidence Intervals for N (Unknown Population Size)}
\usage{
nhgCI_N(
  M,
  m,
  conf_level = 0.95,
  method = "Analog to Clopper-Pearson",
  max_N = 250
)
}
\arguments{
\item{M}{Number of successes in population.}

\item{m}{Fixed number of successes to be observed.}

\item{conf_level}{Confidence level for the interval (default = 0.95).}

\item{method}{Method to compute the interval. Choices are "Analog to Clopper-Pearson", "CMC", "CG", or "Blaker".}

\item{max_N}{Maximum value of N to consider (default = 250).}
}
\value{
A data frame containing the confidence intervals for each observed number of failures observed before the mth success (x).
}
\description{
Wrapper function to compute confidence intervals for the population size (N) in the Negative Hypergeometric Distribution using different methods.
}
\note{
The function assumes \code{m \geq 1} and \code{m \leq M}.
Note that the parameter \code{N} (population size) is unbounded above in the Negative Hypergeometric model.
As a result, the number of observed failures (\code{x}) is theoretically unbounded.
The \code{max_N} parameter sets a practical upper limit for \code{N} during computations.

If the desired value of \code{x} does not appear in the output, try increasing \code{max_N}.
Be aware that increasing \code{max_N} may substantially increase computation time, especially for large \code{m} or when high accuracy is required.
}
\examples{
# Example 1: Using the Analog to Clopper-Pearson method
nhgCI_N(M = 10, m = 5, conf_level = 0.95, method = "Analog to Clopper-Pearson")

# Example 2: Using the CMC method
nhgCI_N(M = 10, m = 5, conf_level = 0.95, method = "CMC")

# Example 3: Using the CG method
nhgCI_N(M = 10, m = 5, conf_level = 0.95, method = "CG")

# Example 4: Using the Blaker method
nhgCI_N(M = 10, m = 5, conf_level = 0.95, method = "Blaker")

}
